+++ src/main/java/tr/com/serkanozal/jillegal/offheap/domain/model/pool/ObjectOffHeapPoolCreateParameter.java
@@ -11,33 +11,33 @@ import tr.com.serkanozal.jillegal.offheap.memory.DirectMemoryService;
 
 public class ObjectOffHeapPoolCreateParameter<T> extends BaseOffHeapPoolCreateParameter<T> {
 
-	protected int objectCount;
+	protected long objectCount;
 	protected ObjectPoolReferenceType referenceType;
 	protected NonPrimitiveFieldAllocationConfigType allocateNonPrimitiveFieldsAtOffHeapConfigType = 
 					NonPrimitiveFieldAllocationConfigType.ONLY_CONFIGURED_NON_PRIMITIVE_FIELDS;
 	
-	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, int objectCount, 
+	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, long objectCount, 
 			ObjectPoolReferenceType referenceType) {
 		super(OffHeapPoolType.OBJECT_POOL, elementType);
 		this.objectCount = objectCount;
 		this.referenceType = referenceType;
 	}
 	
-	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, int objectCount, 
+	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, long objectCount, 
 			ObjectPoolReferenceType referenceType, DirectMemoryService offHeapMemoryService) {
 		super(OffHeapPoolType.OBJECT_POOL, elementType, offHeapMemoryService);
 		this.objectCount = objectCount;
 		this.referenceType = referenceType;
 	}
 	
-	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, int objectCount, 
+	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, long objectCount, 
 			ObjectPoolReferenceType referenceType, boolean makeOffHeapableAsAuto) {
 		super(OffHeapPoolType.OBJECT_POOL, elementType, makeOffHeapableAsAuto);
 		this.objectCount = objectCount;
 		this.referenceType = referenceType;
 	}
 	
-	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, int objectCount, 
+	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, long objectCount, 
 			ObjectPoolReferenceType referenceType, boolean makeOffHeapableAsAuto, 
 			DirectMemoryService offHeapMemoryService) {
 		super(OffHeapPoolType.OBJECT_POOL, elementType, offHeapMemoryService, makeOffHeapableAsAuto);
@@ -45,7 +45,7 @@ public class ObjectOffHeapPoolCreateParameter<T> extends BaseOffHeapPoolCreatePa
 		this.referenceType = referenceType;
 	}
 	
-	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, int objectCount, 
+	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, long objectCount, 
 			ObjectPoolReferenceType referenceType, 
 			NonPrimitiveFieldAllocationConfigType allocateNonPrimitiveFieldsAtOffHeapConfigType, 
 			boolean makeOffHeapableAsAuto) {
@@ -55,7 +55,7 @@ public class ObjectOffHeapPoolCreateParameter<T> extends BaseOffHeapPoolCreatePa
 		this.allocateNonPrimitiveFieldsAtOffHeapConfigType = allocateNonPrimitiveFieldsAtOffHeapConfigType;
 	}
 	
-	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, int objectCount, 
+	public ObjectOffHeapPoolCreateParameter(Class<T> elementType, long objectCount, 
 			ObjectPoolReferenceType referenceType, 
 			NonPrimitiveFieldAllocationConfigType allocateNonPrimitiveFieldsAtOffHeapConfigType, 
 			boolean makeOffHeapableAsAuto, DirectMemoryService offHeapMemoryService) {
@@ -65,11 +65,11 @@ public class ObjectOffHeapPoolCreateParameter<T> extends BaseOffHeapPoolCreatePa
 		this.allocateNonPrimitiveFieldsAtOffHeapConfigType = allocateNonPrimitiveFieldsAtOffHeapConfigType;
 	}
 	
-	public int getObjectCount() {
+	public long getObjectCount() {
 		return objectCount;
 	}
 	
-	public void setObjectCount(int objectCount) {
+	public void setObjectCount(long objectCount) {
 		this.objectCount = objectCount;
 	}
